Added the unit tests (#11)
* Added the unit tests
* Added logout function
* refactor-work

---------

+++ src/main/java/recipeapplication/application/Repository/UserRepository.java
@@ -5,7 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import recipeapplication.application.models.User;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+
+public interface UserRepository extends JpaRepository<User, Integer> {
 
   Optional<User> findByEmail(String email);
